When hash is not found, an event "HashNotFound" is triggered.
This can be use full when, for example, a blog post is book
marked but it no longer features in the recent 10 posts on the
homepage where the plugin is used.

One would rather catch that event and redirect to the desired blog
post rather than show the landing section of the hompage.

Further changes include (mostly related):
* Checks for hash on window.onload
* scrollToAnchor made common function for all hash events/operations
* Few cosmetic changes (brackets, spacing etc.)