Setting isWindowFocused back to true after focus events.
Currently, isWindowFocused is set to true once on initialise(), then either forever true, or changed to false on blur events via the blurHandler(). Having triggered one blur, this is never set back to true upon any kind of focus event triggering on the window.
This, in turn, causes the keyUpHandler to never assign the variable controlPressed back to its original state of false, causing scrolling to not work, due to the MouseWheelHandler checking for !controlPressed.

This commit adds a focusHandler and adds it via window.addEventListener so isWindowFocused gets set to true again.