Fix for #3931 (normalScrollElements option)
There is a fix for the issue #3931, that appears when "normalScrollElements" option is added and mouse cursor leaves from a browser window.

When the mouse cursor left the viewport, the "scrollable" variable was not assigned a value, because the "target" parameter, in this case, is null. So, I just added a check for the "target" parameter value in 2409 line.